requirements.txt
agno
groq
duckduckgo-search
python-dotenv
